fix tests for the rename
